+++ ui/src/app/modules/settings/unpair-single-bridge/unpair-single-bridge.component.ts
@@ -10,7 +10,7 @@ import { firstValueFrom } from 'rxjs'
   templateUrl: './unpair-single-bridge.component.html',
 })
 export class UnpairSingleBridgeComponent implements OnInit, OnDestroy {
-  public pairings: any[]
+  public pairings: any[] = []
   public deleting: null | string = null
   private unpaired = false